requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Face Detection
facenet-pytorch>=2.5.0

# Image/Video Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# ML Utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
